dom/base/test/chrome/test_bug467123.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=467123
-->
<window title="Mozilla Bug 467123"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=467123"
     target="_blank">Mozilla Bug 467123</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

    /** Test for Bug 467123 **/
    var xhr = new XMLHttpRequest();
    xhr.open("GET", "chrome://global/content/bindings/button.xml", false);
    xhr.send();
    ok(xhr.responseXML, "We should have response document!");
    var e = null;
    try {
      var clone = xhr.responseXML.cloneNode(true);
    } catch (ex) {
      e = ex;
    }
    ok(!e, e);
  ]]></script>
</window>
